rag-chatbot/chatbot-backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pypdf2
requests
pandas
langchain
openai
azure-search-documents
jinja2
tiktoken